third_party/blink/web_tests/accessibility/contenteditable-notifications.html

<!DOCTYPE html>
<script src="../resources/testharness.js"></script>
<script src="../resources/testharnessreport.js"></script>
<script src="../resources/run-after-layout-and-paint.js"></script>

<div id="textbox" contenteditable="true">
  First<p>Second</p>
</div>

<script>
'use strict';

async_test((t) => {
  const textbox = document.getElementById('textbox');
  const axTextBox = accessibilityController.accessibleElementById('textbox');
  let valueChangedCount = 0;
  let selectedTextChangedCount = 0;
  const expectedValueChangedIntents = [];
  const expectedSelectedTextChangedIntents = [];

  axTextBox.setNotificationListener(t.step_func((notification, intents) => {
    if (notification == "ValueChanged") {
      assert_array_equals(intents,
                          expectedValueChangedIntents[valueChangedCount]);
      ++valueChangedCount;
    } else if (notification == "SelectedTextChanged") {
      assert_array_equals(intents,
                          expectedSelectedTextChangedIntents[selectedTextChangedCount]);
      ++selectedTextChangedCount;
    }

    if (valueChangedCount >= 4 && selectedTextChangedCount >= 9) {
      t.done();
    }
  }));

  // Initial setting of the selection.
  expectedSelectedTextChangedIntents.push([
    'AXEventIntent(setSelection,character,forward)'
  ]);
  textbox.focus();

  expectedSelectedTextChangedIntents.push([
    'AXEventIntent(moveSelection,lineStart,forward)'
  ]);
  eventSender.keyDown("ArrowDown", []);

  expectedValueChangedIntents.push([
    'AXEventIntent(type,character,forward)'
  ]);
  expectedSelectedTextChangedIntents.push([
    'AXEventIntent(setSelection,character,forward)',
    'AXEventIntent(type,character,forward)'
  ]);
  eventSender.keyDown("w", []);

  expectedSelectedTextChangedIntents.push([
    'AXEventIntent(moveSelection,lineStart,forward)'
  ]);
  eventSender.keyDown("ArrowDown", []);

  expectedValueChangedIntents.push([
    'AXEventIntent(type,character,forward)'
  ]);
  expectedSelectedTextChangedIntents.push([
    'AXEventIntent(type,character,forward)',
    'AXEventIntent(setSelection,character,forward)'
  ]);
  eventSender.keyDown("x", []);

  expectedSelectedTextChangedIntents.push([
    'AXEventIntent(moveSelection,character,backward)'
  ]);
  eventSender.keyDown("ArrowLeft", []);

  expectedValueChangedIntents.push([
    'AXEventIntent(type,character,forward)'
  ]);
  expectedSelectedTextChangedIntents.push([
    'AXEventIntent(setSelection,character,forward)',
    'AXEventIntent(type,character,forward)'
  ]);
  eventSender.keyDown("y", []);

  expectedSelectedTextChangedIntents.push([
    'AXEventIntent(moveSelection,character,backward)'
  ]);
  eventSender.keyDown("ArrowLeft", []);

  expectedValueChangedIntents.push([
    'AXEventIntent(type,character,forward)'
  ]);
  expectedSelectedTextChangedIntents.push([
    'AXEventIntent(setSelection,character,forward)',
    'AXEventIntent(type,character,forward)'
  ]);
  eventSender.keyDown("z", []);
}, 'This test ensures that moving the cursor in a contentEditable sends a selected text change notification, and typing in a contentEditable sends both a value changed and selected text changed notification.');

</script>
